Use alternatatiomn dividers for even-numbered trays
If the number of decks in a tray is even, we can simplify the design
by putting double-width front panels on every other divider, like this:

________________________________
|       |       |       |       |
|       |       |       |       |
|       |       |       |       |
      __|__           __|__

This reduces the number of panels (which makes assembly easier) and
makes each panel twice as wide, which is stronger.

For now, this method is used only when the number of decks is even.
For odd-numbered decks, the original method is used.

I know the code is a bit sloppy.  That will have to get fixed later.